contract-clarity-agent/pyproject.toml
[project]
name = "contract-clarity-agent"
version = "0.1.0"
description = "AI-powered contract analysis assistant using Microsoft Agent Framework"
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "agent-framework[azure]>=0.1.0",
    "streamlit>=1.35.0",
    "firecrawl-py>=1.0.0",
    "docling>=2.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pillow>=10.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.0.0",
    "ruff>=0.3.0",
]
